+++ src/index/table_secondary_index.rs
@@ -1,48 +1,9 @@
-use data_bucket::Link;
-use indexset::concurrent::map::BTreeMap as IndexMap;
-use indexset::concurrent::multimap::BTreeMultiMap as IndexMultiMap;
 use std::collections::HashMap;
-use std::fmt::Debug;
-use std::hash::Hash;
 
+use data_bucket::Link;
+
+use crate::Difference;
 use crate::WorkTableError;
-
-#[derive(Debug)]
-pub struct Difference<AvailableTypes> {
-    pub old: AvailableTypes,
-    pub new: AvailableTypes,
-}
-
-pub trait TableIndex<T> {
-    fn insert(&self, value: T, link: Link) -> Option<Link>;
-    fn remove(&self, value: T, link: Link) -> Option<(T, Link)>;
-}
-
-impl<T> TableIndex<T> for IndexMultiMap<T, Link>
-where
-    T: Eq + Hash + Clone + std::marker::Send + std::cmp::Ord,
-{
-    fn insert(&self, value: T, link: Link) -> Option<Link> {
-        self.insert(value, link)
-    }
-
-    fn remove(&self, value: T, link: Link) -> Option<(T, Link)> {
-        self.remove(&value, &link)
-    }
-}
-
-impl<T> TableIndex<T> for IndexMap<T, Link>
-where
-    T: Eq + Hash + Clone + std::marker::Send + std::cmp::Ord,
-{
-    fn insert(&self, value: T, link: Link) -> Option<Link> {
-        self.insert(value, link)
-    }
-
-    fn remove(&self, value: T, _link: Link) -> Option<(T, Link)> {
-        self.remove(&value)
-    }
-}
 
 pub trait TableSecondaryIndex<Row, AvailableTypes>
 where
